*/scripts: Add tty stdio configuration in deconfig
Unikraft PR #1437 introduced the `LIBPOSIX_TTY_STDIO` config option.
This must be enabled in order to use standard descriptors. In
particular, this enables applications to print to standard output. If
the option is not enabled, nothing is printed. It appears as if the
application is doing nothing or hangs (in case of servers).

Native applications that don't use Musl are excepted from this. They
still work correctly, even if the configuration is not introduced.

Enable the `LIBPOSIX_TTY_STDIO` option in defconfig files.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>